Added decimal place logic for multi asset

+++ leaderboard/party.go
@@ -123,7 +123,7 @@ func hasString(ss []string, s string) bool {
 	return false
 }
 
-func (p *Party) Balance(assetId string, decimalPlaces int64, accountTypes ...string) float64 {
+func (p *Party) Balance(assetId string, decimalPlaces int, accountTypes ...string) float64 {
 	var accu float64
 	for _, acc := range p.Accounts {
 		if acc.Asset.Id == assetId && hasString(accountTypes, acc.Type) {
@@ -143,7 +143,7 @@ func (p *Party) Balance(assetId string, decimalPlaces int64, accountTypes ...str
 	return accu
 }
 
-func (p *Party) CalculateTotalDeposits(asset string, decimalPlaces int64) float64 {
+func (p *Party) CalculateTotalDeposits(asset string, decimalPlaces int) float64 {
 	// Total deposits made in asset
 	var total float64
 	total = 0
